Split EpochCheckpoint and Logger. Avoid too long time for loading keras when importing logger.
